Moved the code for --cli-schema to CliSchema. Updated some nits after in-person discussion. Added version to the root. Shared json serialization options. Created strings for the --cli-schema description.
