+++ src/Cli/dotnet/CliUsage.cs
@@ -36,6 +36,7 @@ sdk-options:
   --list-runtimes   {CliCommandStrings.SDKListRuntimesCommandDefinition}
   --list-sdks       {CliCommandStrings.SDKListSdksCommandDefinition}
   --version         {CliCommandStrings.SDKVersionCommandDefinition}
+  --cli-schema      {CliStrings.SDKSchemaCommandDefinition}
 
 {CliCommandStrings.Commands}:
   build             {CliCommandStrings.BuildDefinition}
